Use fixed-length strings
For compatibility with other libraries we want to use fixed-length
strings. This is difficult since go strings are variable-length and
all the offsets computations in structs are taken directly from go.

The ugly hack to deal with that is to set all strings to fixed
length. By default the value is 20. It can be changed using the
function hdf5.SetStringLength(STRLEN).

Then to write properly the data, the struct has to be defined using a
[STRLEN]byte array instead of a string. To differentiate it from a
regular byte array, the field name has to end in "Str".